Bug 374326 - [Compatibility] ToolControl in toolbar using visibleWhen doesn't show up
Include 3 types of changes:

1. ToolControlContribution doesn't correctly re-fill, because filling is
controled by the @PostConstruct method.  We should uninject the object
on control.dispose and create a new one on fill.

2. push a RAT manager.update(*) outside of the RAT.  As any injected
object can update the IEclipseContext in its @PostConstruct, we need to
be careful

3. remove an MToolControl in an MToolBar from the standard visibility
event handlers.  This is similar to how children of MMenu are already
exempt.

Bug: 374326